Accept hook to display expiration countdown
In order to display the expiration countdown at another hook one
previously had to use the instance of WC_Integration stored in
WC()->integrations.

To make it easier on admins with little coding experience two new
settings were added that allow one to set the hook at which the
expiration countdown.